storage logs added

+++ src/main/java/com/example/shopapplication/Auction.java
@@ -1,6 +1,8 @@
 package com.example.shopapplication;
 
+import java.math.BigDecimal;
 import java.sql.Connection;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.time.LocalDateTime;
@@ -8,21 +10,38 @@ import java.time.LocalDateTime;
 public class Auction {
 
     public void setWinner(Connection connection) throws SQLException {
-
-        // update commodities
         Statement statement = connection.createStatement();
-        String sql = "update allCommodities set number='" + 0 + "', isAuction='" + 0 +
-                "' where isAuction='" + id + "'";
-        statement.executeUpdate(sql);
+        String sql = "select * from allcommodities where isAuction='" + id + "'";
+        ResultSet resultSet = statement.executeQuery(sql);
+        if (resultSet.next()) { // if exists
+            // logs
+            int storageId = resultSet.getInt("storageId");
+            int amount = resultSet.getInt("number");
+            double price = resultSet.getDouble("price");
+            BigDecimal value = new BigDecimal(price).multiply(new BigDecimal(amount));
+            String seller = resultSet.getString("userName");
+            String title = resultSet.getString("title");
+            StorageLog.logAuction(storageId, amount, value.doubleValue(), seller, title, connection);
 
 
-        // delete auction
-        sql = "delete from auction where auctionid='" + id + "'";
-        statement.executeUpdate(sql);
+            // update commodities
+            sql = "update allCommodities set number='" + 0 + "', isAuction='" + 0 +
+                    "' where isAuction='" + id + "'";
+            statement.executeUpdate(sql);
 
-//            // insert into purchases
+            // delete auction
+            sql = "delete from auction where auctionid='" + id + "'";
+            statement.executeUpdate(sql);
+
+//            // todo insert into purchases
 //            sql = "insert into purchases () values ()";
 //            statement.executeQuery(sql);
+
+        } else {
+            System.out.println("Auction.setWinner");
+            System.err.println("not exist!");
+        }
+
 
 
     }
